[SQL] Refactor testing helper classes
