+++ sql-to-dbsp-compiler/SQL-compiler/src/main/java/org/dbsp/sqlCompiler/compiler/frontend/CalciteToDBSPCompiler.java
@@ -1583,4 +1583,8 @@ public class CalciteToDBSPCompiler extends RelVisitor
     public TableContents getTableContents() {
         return this.tableContents;
     }
+
+    public void clearTables() {
+        this.tableContents.clear();
+    }
 }
