+++ sql-to-dbsp-compiler/SQL-compiler/src/main/java/org/dbsp/sqlCompiler/compiler/frontend/TableContents.java
@@ -118,4 +118,13 @@ public class TableContents implements ICompilerComponent {
     public int getTableCount() {
         return this.tablesCreated.size();
     }
+
+    /** Clear the contents of all tables */
+    public void clear() {
+        if (this.tableContents == null)
+            return;
+        for (Map.Entry<String, DBSPZSetLiteral> entry: this.tableContents.entrySet()) {
+            entry.setValue(DBSPZSetLiteral.emptyWithElementType(entry.getValue().getElementType()));
+        }
+    }
 }
