+++ sql-to-dbsp-compiler/SQL-compiler/src/test/java/org/dbsp/sqlCompiler/compiler/sql/suites/TpchTest.java
@@ -17,8 +17,8 @@ public class TpchTest extends BaseSQLTests {
         options.languageOptions.ignoreOrderBy = true;
         options.languageOptions.outputsAreSets = true;
         compiler.compileStatements(tpch);
-        System.err.println(compiler.messages);
-        compiler.throwIfErrorsOccurred();
-        this.addRustTestCase("tpch", compiler, getCircuit(compiler));
+        CompilerCircuitStream ccs = new CompilerCircuitStream(compiler);
+        ccs.showErrors();
+        this.addRustTestCase("tpch", ccs);
     }
 }
